refacto user handle
