+++ src/app/pages/user/user-page/components/user-account-infos/components/user-account-infos-password-change/user-account-infos-password-change.component.html
@@ -1,0 +1,49 @@
+<div class="login-form">
+    <h2 class="pb-3"><i class="bi bi-key"></i> Modification du mot de passe</h2>
+
+    <div class="row g-3">
+        <div class="col-md-12">
+            <app-input-secret
+                    [label]="'Mot de passe actuel'"
+                    [placeholder]="'Entrez votre mot de passe actuel'"
+                    [id]="'currentPasswordInput'"
+                    [required]="true"
+                    [formSent]="formSent"
+                    [inputErrors]="currentPasswordErrors"
+                    (textChange)="currentPassword = $event"
+            ></app-input-secret>
+        </div>
+
+        <div class="col-md-12">
+            <app-input-secret
+                    [label]="'Nouveau mot de passe'"
+                    [placeholder]="'Entrez votre nouveau mot de passe'"
+                    [id]="'newPasswordInput'"
+                    [required]="true"
+                    [formSent]="formSent"
+                    [inputErrors]="newPasswordErrors"
+                    (textChange)="newPassword = $event"
+            ></app-input-secret>
+        </div>
+
+        <div class="col-md-12">
+            <app-input-secret
+                    [label]="'Confirmer le nouveau mot de passe'"
+                    [placeholder]="'Confirmez votre nouveau mot de passe'"
+                    [id]="'confirmNewPasswordInput'"
+                    [required]="true"
+                    [formSent]="formSent"
+                    [inputErrors]="newPasswordConfirmationErrors"
+                    (textChange)="newPasswordConfirmation = $event"
+            ></app-input-secret>
+        </div>
+
+        <div class="col-md-6">
+            <app-send-button
+                    [loading]="loading"
+                    [label]="'Mettre à jour'"
+                    (onClick)="updatePassword()"
+            ></app-send-button>
+        </div>
+    </div>
+</div>
