+++ src/app/pages/user/user-page/components/user-account-infos/components/user-account-infos-password-change/user-account-infos-password-change.component.scss
@@ -1,0 +1,1 @@
+@import "../../user-account-infos.component";
