+++ src/app/pages/user/user-page/components/user-account-infos/components/user-account-infos-personal-data/user-account-infos-personal-data.component.scss
@@ -1,0 +1,5 @@
+@import "../../user-account-infos.component";
+
+.countryErrors {
+  color: #dc3545;
+}
