+++ src/app/pages/user/user-page/components/user-account-infos/user-account-infos.component.html
@@ -1,105 +1,3 @@
-<div class="login-form">
-    <div class="row g-3">
-        <div class="col-md-12">
-            <app-input-text
-                    [label]="'Identifiant'"
-                    [placeholder]="'Entrez votre identifiant'"
-                    [id]="'usernameInput'"
-                    [text]="username"
-                    [required]="true"
-                    [formSent]="formSent"
-                    [inputErrors]="usernameErrors"
-                    (textChange)="username = $event"
-            ></app-input-text>
-        </div>
-
-        <div class="col-md-12">
-            <app-input-email
-                    [formSent]="formSent"
-                    [email]="email"
-                    [emailErrors]="emailErrors"
-                    [isRequired]="true"
-                    (emailChange)="email = $event"
-            ></app-input-email>
-        </div>
-
-        <div class="col-md-12">
-            <label for="validationCountry" class="form-label">Pays</label>
-            <div class="input-group has-validation">
-                <mat-select-country
-                        [required]="true"
-                        [value]="currentCountry"
-                        appearance="fill"
-                        id="validationCountry"
-                        label="Pays"
-                        [itemsLoadSize]="5"
-                        (onCountrySelected)="onCountryChange($event)"
-                        lang="fr">
-                </mat-select-country>
-            </div>
-            <div *ngFor="let error of countryErrors" class="countryErrors">
-                {{ error }}
-            </div>
-        </div>
-
-        <div class="col-md-6">
-            <app-send-button
-                    [loading]="loading"
-                    [label]="'Mettre à jour'"
-                    (onClick)="updateAccount()"
-            ></app-send-button>
-        </div>
-    </div>
-</div>
-
+<app-user-account-infos-personal-data [user]="user"></app-user-account-infos-personal-data>
 <hr>
-
-<div class="login-form">
-    <h2 class="pb-3"><i class="bi bi-key"></i> Modification du mot de passe</h2>
-
-    <div class="row g-3">
-        <div class="col-md-12">
-            <app-input-secret
-                    [label]="'Mot de passe actuel'"
-                    [placeholder]="'Entrez votre mot de passe actuel'"
-                    [id]="'currentPasswordInput'"
-                    [required]="true"
-                    [formSent]="formSent"
-                    [inputErrors]="currentPasswordErrors"
-                    (textChange)="currentPassword = $event"
-            ></app-input-secret>
-        </div>
-
-        <div class="col-md-12">
-            <app-input-secret
-                    [label]="'Nouveau mot de passe'"
-                    [placeholder]="'Entrez votre nouveau mot de passe'"
-                    [id]="'newPasswordInput'"
-                    [required]="true"
-                    [formSent]="formSent"
-                    [inputErrors]="newPasswordErrors"
-                    (textChange)="newPassword = $event"
-            ></app-input-secret>
-        </div>
-
-        <div class="col-md-12">
-            <app-input-secret
-                    [label]="'Confirmer le nouveau mot de passe'"
-                    [placeholder]="'Confirmez votre nouveau mot de passe'"
-                    [id]="'confirmNewPasswordInput'"
-                    [required]="true"
-                    [formSent]="formSent"
-                    [inputErrors]="newPasswordConfirmationErrors"
-                    (textChange)="newPasswordConfirmation = $event"
-            ></app-input-secret>
-        </div>
-
-        <div class="col-md-6">
-            <app-send-button
-                    [loading]="loading"
-                    [label]="'Mettre à jour'"
-                    (onClick)="updatePassword()"
-            ></app-send-button>
-        </div>
-    </div>
-</div>
+<app-user-account-infos-password-change [user]="user"></app-user-account-infos-password-change>
