+++ src/app/pages/user/user-page/components/user-account-infos/user-account-infos.component.scss
@@ -1,6 +1,4 @@
-.countryErrors {
-  color: #dc3545;
-}
+
 
 .login-form {
   color: white;
